moved strategies to new package
