+++ src/main/java/pl/poznan/put/cs/scenariochecker/model/Step.java
@@ -8,4 +8,9 @@ import java.util.List;
 public class Step {
     private String content;
     List<Step> subSteps;
+
+    public Step(String content, List<Step> subSteps) {
+        this.content = content;
+        this.subSteps = subSteps;
+    }
 }
