+++ src/main/java/pl/poznan/put/cs/scenariochecker/transformations/CountStepsScenarioStrategy.java
@@ -1,17 +1,14 @@
-package pl.poznan.put.cs.scenariochecker.controller.strategy;
+package pl.poznan.put.cs.scenariochecker.transformations;
 
 import pl.poznan.put.cs.scenariochecker.model.Scenario;
 import pl.poznan.put.cs.scenariochecker.model.Step;
 
 import java.util.List;
 
-import static pl.poznan.put.cs.scenariochecker.controller.strategy.ScenarioHelper.isSpecialStep;
-
 public class CountStepsScenarioStrategy extends ScenarioStrategy {
 
     /**
-     * @param scenario
-     * This method is responsible for proccessing scenario when user want to count number of steps
+     * @param scenario This method is responsible for proccessing scenario when user want to count number of steps
      */
     @Override
     public void processScenario(Scenario scenario) {
@@ -24,13 +21,10 @@ public class CountStepsScenarioStrategy extends ScenarioStrategy {
      */
     private int countSteps(List<Step> steps) {
         long deepCount = steps.stream()
-                .filter(step -> isSpecialStep(step))
+                .filter(step -> ScenarioHelper.isSpecialStep(step))
                 .map(step -> countSteps(step.getSubSteps()))
                 .mapToLong(Integer::intValue)
                 .sum();
         return (int) (steps.size() + deepCount);
     }
-
-
-
 }
